Debugged get_disk_storage() as well as added info struct and filler function for said struct

+++ src/main.c
@@ -16,8 +16,8 @@ int main() {
   }
 
 
-  int version = get_disk_storage();
-  
-  printf("\n%d\n\n", version);
-  //free(version);  
+  double version = get_disk_storage();
+
+  printf("\n%f\n\n", version);
+  //free(version);
 }
